🚚 Move: GatheringDocumentProvider 위치 이동
