+++ src/test/java/com/withus/withmebe/search/controller/GatheringDocumentControllerTest.java
@@ -1,6 +1,6 @@
 package com.withus.withmebe.search.controller;
 
-import static com.withus.withmebe.search.util.GatheringDocumentProvider.getStubbedGatheringDocument;
+import static util.StubbedObjectProvider.GatheringDocumentProvider.getStubbedGatheringDocument;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.BDDMockito.given;
